+++ R/data.R
@@ -23,10 +23,11 @@
 #'   * **SRS**: Spatial Reference System (CRS) value, identified by the
 #'     corresponding [EPSG](https://en.wikipedia.org/wiki/EPSG_Geodetic_Parameter_Dataset)
 #'     code.
-#'   * **Description**: Description of the SRS/EPSG code.
-#'   * **ovc_service**: Logical. Is this code valid on OVC services?
-#'   * **wfs_service**: Logical. Is this code valid on INSPIRE WFS services?
-#'
+#' \describe{
+#'   \item{Description}{Description of the SRS/EPSG code.}
+#'   \item{ovc_service}{Logical. Is this code valid on OVC services?}
+#'   \item{wfs_service}{Logical. Is this code valid on INSPIRE WFS services?}
+#' }
 #' @details
 #'
 #' ```{r, echo=FALSE}