+++ tests/testthat/test-catr_ovc_rccoor_distancia.R
@@ -14,7 +14,10 @@ test_that("return tibble given SRS", {
   skip_if_offline()
   skip_on_os("linux")
 
-  result <- catr_ovc_get_rccoor_distancia(lat = 40.963200, lon = -5.671420, "4326")
+  result <- catr_ovc_get_rccoor_distancia(
+    lat = 40.963200,
+    lon = -5.671420, "4326"
+  )
   expect_s3_class(result, "tbl")
   expect_true(is.numeric(result$geo.xcen))
   expect_true(is.numeric(result$geo.ycen))
@@ -45,7 +48,10 @@ test_that("check fields given SRS", {
   skip_if_offline()
   skip_on_os("linux")
 
-  result <- catr_ovc_get_rccoor_distancia(lat = 40.963200, lon = -5.671420, 4230)
+  result <- catr_ovc_get_rccoor_distancia(
+    lat = 40.963200,
+    lon = -5.671420, 4230
+  )
   expect_true((is.character(result$address) & is.character(result$refcat) &
     is.character(result$cmun_ine)))
 })
@@ -55,7 +61,10 @@ test_that("if data is known return a tibble with 3 cols", {
   skip_if_offline()
   skip_on_os("linux")
 
-  expect_message(catr_ovc_get_rccoor_distancia(lat = 99999999, lon = -999999999))
+  expect_message(catr_ovc_get_rccoor_distancia(
+    lat = 99999999,
+    lon = -999999999
+  ))
   result <- catr_ovc_get_rccoor_distancia(lat = 99999999, lon = -999999999)
   expect_true(ncol(result) == 3)
 })