+++ NEWS.md
@@ -1,6 +1,7 @@
 # CatastRo (development version)
 
 -   Housekeeping and update of documentation.
+-   Adapt `catr_wms_get_layer()` to **mapSpain** (\>= 0.7.0).
 
 # CatastRo 0.2.2